raspberry-weigend/requirements.txt
py-remote-deployer